cmd: remove fetch/process cmds
Simplify to just trimming + compressing. The other parts are a thin
wrapper around the acvptool and haven't been very helpful in practice.
