+++ go.mod
@@ -2,4 +2,4 @@ module github.com/cpu/go-acvp
 
 go 1.22.5
 
-require github.com/dsnet/compress v0.0.1 // indirect
+require github.com/dsnet/compress v0.0.1
